move miner types to `unsigned`
