+++ frame/election-provider-multi-phase/src/lib.rs
@@ -268,22 +268,6 @@ pub type CompactAccuracyOf<T> = <CompactOf<T> as CompactSolution>::Accuracy;
 /// The accuracy of the election, when computed on-chain. Equal to [`Config::OnChainAccuracy`].
 pub type OnChainAccuracyOf<T> = <T as Config>::OnChainAccuracy;
 
-/// A voter's fundamental data: their ID, their stake, and the list of candidates for whom they voted.
-pub type Voter<T> = (
-	<T as frame_system::Config>::AccountId,
-	sp_npos_elections::VoteWeight,
-	Vec<<T as frame_system::Config>::AccountId>,
-);
-
-/// The relative distribution of a voter's stake among the winning targets.
-pub type Assignment<T> = sp_npos_elections::Assignment<
-	<T as frame_system::Config>::AccountId,
-	CompactAccuracyOf<T>,
->;
-
-/// The [`IndexAssignment`] type specialized for a particular runtime `T`.
-pub type IndexAssignmentOf<T> = sp_npos_elections::IndexAssignmentOf<CompactOf<T>>;
-
 /// Wrapper type that implements the configurations needed for the on-chain backup.
 struct OnChainConfig<T: Config>(sp_std::marker::PhantomData<T>);
 impl<T: Config> onchain::Config for OnChainConfig<T> {
@@ -573,7 +557,7 @@ pub mod pallet {
 			+ Eq
 			+ Clone
 			+ sp_std::fmt::Debug
-			+ for<'a> sp_std::convert::TryFrom<&'a [IndexAssignmentOf<Self>], Error = sp_npos_elections::Error>
+			+ for<'a> sp_std::convert::TryFrom<&'a [crate::unsigned::IndexAssignmentOf<Self>], Error = sp_npos_elections::Error>
 			+ CompactSolution;
 
 		/// Accuracy used for fallback on-chain election.
